Save Scene 2:
Created the apple by making a sphere and coloring it red, similar to the apple tree. The apple has gravity, unlike the apple tree which does not. Created a tag for the apple called "Apple". Made apple into a prefab. Created a basket using the same method as apple, also made it into a prefab.