Save Scene 4:
Added more code in VS to position and add movement to the AppleTree. Set the speed to 10. AppleTree moves back and forth on the top of the scene. Changed the code so that Appletree will change direction an average of 1 time every second in FixedUpdate(). Added Apple to Apple prefab in AppleTree Inspector. Added to script for the apples to drop.