Save Scene 4:
Gave the Slingshot a halo effect. Wrote code to show when the slingshot is active. Also wrote code for the game to detect when the mouse is close enough to the Slingshot to interact with it and to instantiate a projectile when the user clicks the mouse near the slingshot. Assigned the projectile to the slingshot.