Save Scene 9:
I got all the goals to work so all the mazes should be playable. I just need to make it so that when the player sees the "You Win" text, it will destroy the maze and instantiate the next maze. Also I don't have time for the Bullet_Rain idea so there will be no obstacles for now.